// src/models/history.ts
export type Historial = {
    id_historial_medico: number;
    id_paciente: number;
    id_tipo_sangre: number;
    id_usuario: number;
    antecedentes_heredofamiliares: string;
    alergias: string;
    patologias: string;
    intervencion_quirurgica: number;
    transfucion_sanguinea: number;
    donacion_sanguinea: number;
    observaciones: string;
    fecha_registro: string;
};